src/pyproject.toml
[tool.poetry]
name = "semanticsearch"
version = "0.1.0"
description = ""
authors = ["Carlos Martín Ruiz <crlsmartnruiz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.2.1"
sentence-transformers = "^3.0.0"
langchain-huggingface = "^0.0.1"
elasticsearch = "8.14.0"
pymongo = {extras = ["srv"], version = "^4.7.3"}
python-dotenv = "^1.0.1"
openai = "^1.31.2"
langchain-elasticsearch = "^0.2.2"
langchain-openai = "^0.1.8"
tiktoken = "^0.7.0"
pypdf = "^4.2.0"
langchain-community = "^0.2.4"
langchainhub = "^0.1.18"
chromadb = "^0.5.2"
llama-cpp-python = "^0.2.78"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
